requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
asyncpg==0.30.0
annotated-types==0.7.0
attrs==25.4.0
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
librt==0.7.3
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
mypy==1.19.0
mypy_extensions==1.1.0
numpy==2.3.5
packaging==25.0
pathspec==0.12.1
pgvector==0.4.2
platformdirs==4.5.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pytokens==0.3.0
openai==1.63.0
ruff==0.14.8
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
win32_setctime==1.2.0
yarl==1.22.0
